Separate ideas of discrete randvars and root discrete randvars
