+++ alea/discrete/__init__.py
@@ -1,2 +1,3 @@
 from .discrete_randvar import *
+from .root_randvar import *
 from .special_randvars import *
