+++ tests/test_ops_discrete_randvars.py
@@ -4,7 +4,7 @@ from alea.discrete import BinomialRandVar
 
 
 def almost_equal(x, y):
-    return abs(x - y) <= 1e-2
+    return abs(x - y) <= 1e-5
 
 
 class TestConstantSum:
@@ -100,7 +100,7 @@ class TestDiscreteMultiply:
         Z = X * (X * (X + Y))
         Z2 = (X ** 3) + (X ** 2) * Y
         assert(almost_equal(Z.mean(), Z2.mean()))
-        assert(almost_equal(Z.mean(), Z2.sample_average()))
+        # assert(almost_equal(Z.variance(), Z2.variance()))
 
 
     def test_foil(self):
@@ -109,6 +109,7 @@ class TestDiscreteMultiply:
         Z = (X + Y) * (X + Y)
         Z2 = (X * X) + (X * Y) + (X * Y) + (Y * Y)
         assert(almost_equal(Z.mean(), Z2.mean()))
+        # assert(almost_equal(Z.variance(), Z2.variance()))
 
 
     def test_slow_exponentiation(self):
@@ -117,3 +118,4 @@ class TestDiscreteMultiply:
         Z = X * X * X * X * X * X
         Z2 = X ** 6
         assert(almost_equal(Z.mean(), Z2.mean()))
+        # assert(almost_equal(Z.variance(), Z2.variance()))
